add new ProductCard template and use in Shop and Cart
ProductCard allows for conditional content using the @Prop magic.

deleteFromCart is refactored out of Cart  and now cart simple handles the "deleted" event caught by v-on:"deleted" ... this follows the pattern in Modal.vue with `successHandler` and `cancelHandler` except the new function is called `deleteHandler`

so the deleteFromCart function is inside ProductCard because it makes sense that a vue file that is simple concerne with 1 product could easily handle the store.state.cart for that 1 product. Smaller scope and separation of concerns.

same for the addToCart function, refactored out of ProductList and into ProductCard. some reason as above.

also added images to public folder that can be referenced through typescript

template code for Shop and Cart are now much shorter, more consise, more readable

also cleaned up some unused code in Home