Shop now knows when a filter is checked
I am going to avoid using this.$nextTick because Vue has "best practice" that says `Although Vue.js generally encourages developers to think in a “data-driven” way and avoid touching the DOM directly`

https://012.vuejs.org/guide/best-practices.html#Understanding_Async_Updates

That will also free the code from the mess of having to handle the `observer` object. And leave that implementation detail to Vue by simply using the `v-for` syntax and the data that is availible within that block of code.

And it turns out that `v-on:change` is the way to handle a checkbox change event which happens after the `v-on:click` actually. I wish I knew that sooner.

also ran npm `run lint --fix` on Prodcut vue files