create array.utils.ts to extend Array with union and intersection
So I have been struggling for the past few commits because the union and intersection code that I added using "Set" did not work as expected for objects. Now using JSON.stringify and JSON.parse it does work with objects.

I think the runtime would be in the order of n^2 for intersection because of the `setArr.has(x)` unless Set is implemented with a Map which could mean that intersection is as fast as O(n)

I think -- if I counted right -- union woud run on the order of (5*n + 5m) where n is the length of `this` and m is the length of `arr`
* maybe one way to speed things up  by a little bit would be to have `setThis = new Set()` on line 1 and in the `this.map` for `strEachThis` the stringified value could be immediately added `setThis.add(stringifiedValue)` ... so actually in that case use `this.forEach` rather than `this.map`